requirements.txt
pytest
pylint
ipdb
jupyterlab
numpy
pandas
matplotlib
seaborn
scikit-learn
google-cloud-storage
colorama
